# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HASN Digital Services Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List services
      parameters:
        - in: query
          name: categoryId
          schema:
            type: integer
        - in: query
          name: featured
          schema:
            type: boolean
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Services
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceList"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
...
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get service by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Providers
  /providers:
    get:
      operationId: listProviders
      tags: [providers]
      summary: List service providers
      responses:
        "200":
          description: Providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Provider"
    post:
      operationId: createProvider
      tags: [providers]
      summary: Add a new service provider
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
  /providers/{id}:
    get:
      operationId: getProvider
      tags: [providers]
      summary: Get provider by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
    patch:
      operationId: updateProvider
      tags: [providers]
      summary: Update provider
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProviderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
    delete:
      operationId: deleteProvider
      tags: [providers]
      summary: Delete provider
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /providers/{id}/services:
    get:
      operationId: listProviderServices
      tags: [providers]
      summary: List all services for a specific provider
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Provider services
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceList"
    patch:
      operationId: updateProviderServicesVisibility
      tags: [providers]
      summary: Bulk update visibility for provider services
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkVisibilityInput"
      responses:
        "200":
          description: Updated count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdatedCount"
  /providers/{id}/sync:
    post:
      operationId: syncProviderServices
      tags: [providers]
      summary: Sync services from provider API
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Synced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResult"

  # Wallet
...
components:
  schemas:
...
    Service:
      type: object
      required: [id, name, price, isVisible, isFeatured, categoryId]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        isVisible:
          type: boolean
        isFeatured:
          type: boolean
        categoryId:
          type: ["integer", "null"]
        providerId:
          type: ["integer", "null"]
        providerServiceId:
          type: ["string", "null"]
        profitMargin:
          type: ["number", "null"]
    ServiceList:
      type: object
      required: [services, total, page, limit]
      properties:
        services:
          type: array
          items:
            $ref: "#/components/schemas/Service"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ServiceInput:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        price:
          type: number
        categoryId:
          type: integer
        providerId:
          type: integer
        providerServiceId:
          type: string
        isFeatured:
          type: boolean
        isVisible:
          type: boolean
        profitMargin:
          type: number
    ServiceUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        price:
          type: number
        categoryId:
          type: integer
        providerId:
          type: integer
        providerServiceId:
          type: string
        isFeatured:
          type: boolean
        isVisible:
          type: boolean
        profitMargin:
          type: number
...
    BulkVisibilityInput:
      type: object
      required: [serviceIds, isVisible]
      properties:
        serviceIds:
          type: array
          items:
            type: integer
        isVisible:
          type: boolean
    UpdatedCount:
      type: object
      required: [updated]
      properties:
        updated:
          type: integer

    # Provider
    Provider:
      type: object
      required: [id, name, apiUrl, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        apiUrl:
          type: string
        isActive:
          type: boolean
        defaultProfitMargin:
          type: ["number", "null"]
        defaultCategoryId:
          type: ["integer", "null"]
        lastSyncedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    ProviderInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        apiUrl:
          type: string
        apiKey:
          type: string
        defaultProfitMargin:
          type: number
        defaultCategoryId:
          type: integer
    ProviderUpdate:
      type: object
      properties:
        name:
          type: string
        apiUrl:
          type: string
        apiKey:
          type: string
        defaultProfitMargin:
          type: number
        defaultCategoryId:
          type: ["integer", "null"]
        isActive:
          type: boolean
    SyncResult:
      type: object
      required: [synced, total]
      properties:
        synced:
          type: integer
        total:
          type: integer
        message:
          type: string

    # Wallet
...